# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "AAStar BackEndDashBoard API",
        "contact": {
            "name": "AAStar BackEndDashBoard",
            "url": "https://aastar.xyz"
        },
        "version": "v1"
    },
    "basePath": "/api",
    "paths": {
        "/api/healthz": {
            "get": {
                "description": "Get Healthz",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Healthz"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/v1/api_key": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "GetApiKey",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "GetApiKey"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Api Key",
                        "name": "api_key",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "UpdateApiKey",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "UpdateApiKey"
                ],
                "parameters": [
                    {
                        "description": "UploadApiKeyRequest Model",
                        "name": "uploadApiKeyRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UploadApiKeyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "DeleteApiKey",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "DeleteApiKey"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Api Key",
                        "name": "api_key",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/v1/api_key/apply": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "ApplyApiKey",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "ApplyApiKey"
                ],
                "parameters": [
                    {
                        "description": "UploadApiKeyRequest Model",
                        "name": "applyApiKeyRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ApplyApiKeyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/v1/api_key/list": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "GetApiKeyList",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "GetApiKeyList"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/v1/paymaster_strategy": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "GetStrategy",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "GetStrategy"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "PaymasterStrategy Code",
                        "name": "strategy_code",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "UpdateStrategy",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "UpdateStrategy"
                ],
                "parameters": [
                    {
                        "description": "UploadStrategyRequest Model",
                        "name": "uploadStrategyRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UploadStrategyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "AddStrategy",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "AddStrategy"
                ],
                "parameters": [
                    {
                        "description": "UploadStrategyRequest Model",
                        "name": "uploadStrategyRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UploadStrategyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "DeleteStrategy",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "DeleteStrategy"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "PaymasterStrategy Code",
                        "name": "strategy_code",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/v1/paymaster_strategy/list": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "GetStrategyList",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "GetStrategyList"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/v1/user": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "GetUserInfo",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "GetUserInfo"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/oauth/github": {
            "get": {
                "description": "Github OAuth Login",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "OAuth"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Github OAuth Code",
                        "name": "code",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/oauth/password": {
            "post": {
                "description": "PasswordOauthLogin",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Oauth"
                ],
                "parameters": [
                    {
                        "description": "PasswordRequest Model",
                        "name": "passwordRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/oauth.PasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "definitions": {
        "model.ApplyApiKeyRequest": {
            "type": "object",
            "properties": {
                "api_key_name": {
                    "type": "string"
                }
            }
        },
        "model.UploadApiKeyRequest": {
            "type": "object",
            "properties": {
                "api_key": {
                    "type": "string"
                },
                "api_key_name": {
                    "type": "string"
                }
            }
        },
        "model.UploadStrategyRequest": {
            "type": "object",
            "properties": {
                "address_block_list": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "chain_id_whitelist": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "day_max_usd": {
                    "type": "number"
                },
                "description": {
                    "type": "string"
                },
                "end_time": {
                    "type": "integer"
                },
                "global_max_usd": {
                    "type": "number"
                },
                "per_user_max_usd": {
                    "type": "number"
                },
                "project_code": {
                    "type": "string"
                },
                "start_time": {
                    "type": "integer"
                },
                "strategy_code": {
                    "type": "string"
                },
                "strategy_name": {
                    "type": "string"
                }
            }
        },
        "oauth.PasswordRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "JWT": {
            "description": "Type 'Bearer \\\u003cTOKEN\\\u003e' to correctly set the AccessToken",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}